Refaktoryzacja w kinematiclib DataChannelCollection Porzadki w kodzie Eventy
git-svn-id: https://localhost:8443/svn/hm-mde-1@508 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ trunk/src/edr/include/core/ObjectWrapperCollection.h
@@ -145,7 +145,7 @@ public:
     const ObjectWrapperConstPtr & getObject(int index) const
     {
         // wyjatek zamiast asercji (na wypadek trybu release)
-        if (!(index < size() && index >= 0)) {
+        if (!(index < static_cast<int>(size()) && index >= 0)) {
             throw std::runtime_error("ObjectWrapperCollection::getObject - wrong index");
         }
